% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{agg.by.region}
\alias{agg.by.region}
\title{Aggregate the named values in gr by region using the aggregation function.}
\usage{
agg.by.region(gr, regions, value.col, aggregation.fn = max)
}
\description{
Aggregate the named values in gr by region using the aggregation function.
}

