% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cell.names}
\alias{cell.names}
\title{Assume all columns that are not named cell, chrom, start, stop, end are cell names}
\usage{
cell.names(df)
}
\description{
Assume all columns that are not named cell, chrom, start, stop, end are cell names
}

