% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQB.R
\name{fitQB}
\alias{fitQB}
\title{Function to fit quasi-binomial models}
\usage{
fitQB(
  countData,
  tx2gene,
  design,
  speed = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam(),
  verbose = TRUE
)
}
\arguments{
\item{countData}{A `DataFrame` or `matrix` of transcript expression values.
The transcripts are along the rows and samples/cells along the columns.}

\item{tx2gene}{A `DataFrame` or `matrix` that links each transcript in
the countData object to its corresponding gene ID.}

\item{design}{A `DataFrame` with information on the design. It has
the same number of rows as the number of columns (samples/cells) of
`countData`.}

\item{speed}{`boolean(1)` Logical, defaults to FALSE. Set to TRUE if
the quasi-binomial models should be fitted with the \code{speedglm} package
rather than the classical glm procedure, which depending on the dataset might
reduce computational time and memory usage. See the \code{speedglm} package
for details.}

\item{parallel}{`boolean(1)` Logical, defaults to FALSE. Set to TRUE if to
parallellize the fitting procedure.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{bpparam} in \code{BiocParallel} package for details.}

\item{verbose}{`boolean(1)` Logical, defaults to TRUE. Used to indicate if a
progress bar should be printed during the fitting procedure.}
}
\value{
A list of objects of the `StatModel` class.
}
\description{
Parameter estimation of quasi-binomial models.
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
