% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ladder.to.test.idx}
\alias{ladder.to.test.idx}
\title{The index of the region in the test regions, given its index in the ladder regions}
\usage{
ladder.to.test.idx(idx)
}
\description{
The index of the region in the test regions, given its index in the ladder regions
}

