% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{load.chip.features}
\alias{load.chip.features}
\title{Load all the ChIP features for the TF and return as a DataFrame}
\usage{
load.chip.features(tf)
}
\description{
Load all the ChIP features for the TF and return as a DataFrame
}

