% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{load.motif.dir}
\alias{load.motif.dir}
\title{Load motif scan results from directory}
\usage{
load.motif.dir(motifs.dir, prior.log.odds = .PRIOR.LOG.ODDS,
  maximum.BF = 10)
}
\description{
Load motif scan results from directory
}

