% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{load.motif.scan}
\alias{load.motif.scan}
\title{Load motif scan results}
\usage{
load.motif.scan(results.path, seqs.path, prior.log.odds = .PRIOR.LOG.ODDS,
  maximum.BF = 10)
}
\description{
Load motif scan results
}

