% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDTU.R
\name{plotDTU}
\alias{plotDTU}
\title{Plot function to visualize differential transcript usage (DTU)}
\usage{
plotDTU(
  topTable,
  contrast,
  metaData,
  countData,
  tx2gene,
  transcripts = NULL,
  genes = NULL,
  top.n = 6
)
}
\arguments{
\item{topTable}{A `DataFrame` displaying the significance of DTU for each
transcript as generated by the `topTable` function from this `qbDTU` package.
If `topTable` is a `List` of `DataFrames`, one should specify (index) the
`DataFrame` of interest.}

\item{metaData}{A `DataFrame` that has information on the metaData (i.e. cell type,
cluster ID, tissue type, ...) of the experiment.}

\item{countData}{A `DataFrame` or `matrix` of transcript expression values.
The transcripts are along the rows and samples/cells along the columns.}

\item{tx2gene}{A `DataFrame` or `matrix` that links each transcript in
the countData object to its corresponding gene ID.}

\item{transcripts}{A `character` or `character vector` of transcript IDs,
to specify which transcripts should be visualized. Can be used together with
`genes`. If not specified, `plotDTU` will check if the `genes` slot is
specified.}

\item{genes}{A single `character` or `character vector` of gene IDs,
to specify the genes for which the individual transcripts should be visualized.
Can be used together with `transcripts`. If not specified, `plotDTU` will check
if the `transcripts` slot is specified.}

\item{top.n}{A `numeric` value. If neither `transcripts` nor `genes` was
was specified, this argument leads to the visualization of the `n` most
significant DTU transcripts in the contrast. Defaults to 6 transcripts}
}
\value{
A plot directly displayed in the current R session.
}
\description{
Plot function to visualize differential transcript usage (DTU)
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
