% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbDTU-framework.R
\docType{class}
\name{StatModel-class}
\alias{StatModel-class}
\alias{.StatModel}
\title{The StatModel class for qbDTU}
\description{
classes for qbDTU
}
\section{Slots}{

\describe{
\item{\code{type}}{The type of the used model}

\item{\code{params}}{A list containing information of the used model}

\item{\code{varPosterior}}{A vector of posterior variance}

\item{\code{dfPosterior}}{A vector of posterior degrees of freedom}
}}

\author{
Jeroen Gilis
}
