% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{read.chip.labels}
\alias{read.chip.labels}
\title{Read the ChIP binding labels for the TF into a S4Vectors::DataFrame}
\usage{
read.chip.labels(tf)
}
\description{
Read the ChIP binding labels for the TF into a S4Vectors::DataFrame
}

