% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Rle.from.sparse}
\alias{Rle.from.sparse}
\title{Construct a run-length encoded vector with zeros everywhere else than specified}
\usage{
Rle.from.sparse(len, idxs, x)
}
\description{
Construct a run-length encoded vector with zeros everywhere else than specified
}

