% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_disconnect}
\alias{sat_disconnect}
\title{Saturn Disconnect}
\usage{
sat_disconnect(session)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}
}
\description{
Saturn Disconnect
}
\examples{
\dontrun{
session <- sat_connect()
sat_disconnect(session)
}

}
\seealso{
\code{\link[ssh]{ssh}}
}
