% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_execute}
\alias{sat_execute}
\title{Saturn Execute}
\usage{
sat_execute(session, command = "whoami", error = TRUE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{command}{The command or script to execute}

\item{error}{automatically raise an error if the exit status is non-zero}
}
\value{
list containing exit status, buffered raw stdout, and buffered raw
stderr
}
\description{
Saturn Execute
}
\examples{
\dontrun{
session <- sat_connect()
out <- sat_execute(session)
rawToChar(out$stdout)
}

}
\seealso{
\code{\link[ssh]{ssh_exec}}
}
