% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_edit_cron}
\alias{sat_job_edit_cron}
\title{Saturn Job Edit Cron}
\usage{
sat_job_edit_cron(job_name)
}
\arguments{
\item{job_name}{job name}
}
\description{
Saturn Job Edit Cron
}
\note{
The output of the cron command will be automatically sent to your
local email account. If you want to stop receiving these emails, you can add
>/dev/null 2>&1 to the syntax as in the following example: 0 5 * * *
/root/backup.sh >/dev/null 2>&1
}
\examples{
\dontrun{
job_name <- "my_job"
sat_job_init(job_name)
sat_job_edit_cron(job_name)
}

}
