% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_edit_source}
\alias{sat_job_edit_source}
\title{Saturn Job Edit Source}
\usage{
sat_job_edit_source(job_name)
}
\arguments{
\item{job_name}{job name}
}
\description{
Saturn Job Edit Source
}
\examples{
\dontrun{
job_name <- "my_job"
sat_job_init(job_name, edit_script = FALSE)
sat_job_edit_source(job_name)
}
}
