% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_running}
\alias{sat_job_running}
\title{Saturn Job Running}
\usage{
sat_job_running(session, job_name)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{job_name}{job name}
}
\value{
logical indicating if job is running
}
\description{
Saturn Job Running
}
\examples{
\dontrun{
session <- sat_connect()
job_name <- "my_job"
sat_job_init(job_name)
sat_job_upload(session, job_name)
sat_job_submit(session, job_name)
sat_job_running(session, job_name)
}
}
