% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_schedule}
\alias{sat_job_schedule}
\title{Saturn Job Schedule}
\usage{
sat_job_schedule(session, job_name, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{job_name}{job name}

\item{...}{additional arguments passed to \code{\link{sat_execute}},
\code{\link{sat_upload}}}
}
\description{
Saturn Job Schedule
}
\details{
\tabular{ll}{
\strong{Command}         \tab \strong{Details}               \cr
crontab -e               \tab create and edit a crontab file \cr
crontab -l               \tab view contents of crontab file  \cr
crontab -r               \tab unschedule all jobs            \cr
crontab -a job_name.cron \tab schedule a job                 \cr
}
}
\note{
check mail with mail command
}
\examples{
\dontrun{
session <- sat_connect()
job_name <- "my_job"
sat_job_init(job_name, add_cron = TRUE)
sat_job_upload(session, job_name)
sat_job_schedule(session, job_name)
}

}
