% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_submit}
\alias{sat_job_submit}
\title{Saturn Job Submit}
\usage{
sat_job_submit(session, job_name, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{job_name}{job name}

\item{...}{additional arguments passed to \code{\link{sat_execute}},
\code{\link{sat_upload}}}
}
\description{
Saturn Job Submit
}
\examples{
\dontrun{
session <- sat_connect()
job_name <- "my_job"
sat_job_init(job_name)
sat_job_upload(session, job_name)
sat_job_submit(session, job_name)
}
}
