% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_job_upload}
\alias{sat_job_upload}
\title{Saturn Job Upload}
\usage{
sat_job_upload(session, job_name, mk_spath = FALSE, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{job_name}{job name}

\item{mk_spath}{make the spath if it does not exist}

\item{...}{additional arguments passed to \code{\link{sat_execute}} and
\code{\link{sat_upload}}}
}
\description{
Saturn Job Upload
}
\examples{
\dontrun{
session <- sat_connect()
job_name <- "my_job"
sat_job_init(job_name)
sat_job_upload(session, job_name)
}
}
