% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_pbsnodes}
\alias{sat_pbsnodes}
\title{Saturn pbsnodes}
\usage{
sat_pbsnodes(session, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{...}{additional arguments passed to \code{\link{sat_qstat}}}
}
\value{
status of all compute nodes
}
\description{
Saturn pbsnodes
}
\examples{
\dontrun{
session <- sat_connect()
r <- sat_pbsnodes(session)
cat(rawToChar(r$stdout))
}
}
