% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat.R
\name{sat_qstat}
\alias{sat_qstat}
\title{Saturn qstat Command}
\usage{
sat_qstat(session, args, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{sat_connect}}}

\item{args}{arguments to pass to qstat}

\item{...}{additional arguments passed to \code{\link{sat_execute}}}
}
\description{
Saturn qstat Command
}
\details{
\tabular{ll}{
\strong{Argument} \tab \strong{Details}              \cr
-q        \tab list all queues                       \cr
-a        \tab list all jobs                         \cr
-u userid \tab list jobs for userid                  \cr
-r        \tab list running jobs                     \cr
-f job_id \tab list full information about job_id    \cr
-Qf queue \tab list full information about queue     \cr
-B        \tab list summary status of the job server \cr
}
}
\examples{
\dontrun{
session <- sat_connect()
r <- sat_qstat(session, "-a")
cat(rawToChar(r$stdout))
}

}
