% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn}
\alias{saturn}
\title{A function to draw ASCII of the planet Saturn}
\usage{
saturn(show = TRUE)
}
\arguments{
\item{show}{it shows an image of the Saturn if default or TRUE}
}
\value{
satrun ASCII
}
\description{
A function to draw ASCII of the planet Saturn
}
