% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sparse.to.rle}
\alias{sparse.to.rle}
\title{Construct a run-length encoded vector with zeros everywhere else than specified}
\usage{
sparse.to.rle(len, idxs, x)
}
\description{
DEPRECATED: Rle.from.sparse is much more efficient.
}

