% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satuRn-framework.R
\name{StatModel}
\alias{StatModel}
\title{StatModel}
\usage{
StatModel(
  type = "fitError",
  params = list(),
  varPosterior = is.numeric(NA),
  dfPosterior = is.numeric(NA)
)
}
\arguments{
\item{type}{default set to fiterror, can be a glm}

\item{params}{A list containing the parameters of the fitted glm}

\item{varPosterior}{Numeric, posterior variance of the glm, default is NA}

\item{dfPosterior}{Numeric, posterior degrees of freedom of the glm, default is NA}
}
\description{
Function for contstructing a new `StatModel` object.
}
