% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumExp_vignette}
\alias{sumExp_vignette}
\title{A summarizedExperiment derived from Tasic case study (explain how).}
\format{
An object of class \code{SummarizedExperiment} with 13042 rows and 60 columns.
}
\usage{
data(sumExp_vignette)
}
\description{
A summarizedExperiment derived from Tasic case study (explain how).
}
\keyword{datasets}
