% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{summarise.chip.peaks}
\alias{summarise.chip.peaks}
\title{Summarise ChIP peaks by test regions}
\usage{
summarise.chip.peaks(cell, tf, type = "conservative", aggregation.fn = max)
}
\description{
Summarise ChIP peaks by test regions
}

