% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnase.R
\name{summarise.dnase}
\alias{summarise.dnase}
\title{Summarise DNase peaks by ChIP-regions}
\usage{
summarise.dnase(cell, type = "conservative", aggregation.fn = max)
}
\description{
Summarise DNase peaks by ChIP-regions
}

