% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tasic_counts_vignette}
\alias{Tasic_counts_vignette}
\title{A DataFrame with counts derived from Tasic case study (explain how).}
\format{
An object of class \code{matrix} with 35423 rows and 150 columns.
}
\usage{
data(Tasic_counts_vignette)
}
\description{
A DataFrame with counts derived from Tasic case study (explain how).
}
\keyword{datasets}
