% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tasic_metadata_vignette}
\alias{Tasic_metadata_vignette}
\title{A DataFrame with metadata derived from Tasic case study (explain how).}
\format{
An object of class \code{data.frame} with 60 rows and 3 columns.
}
\usage{
data(Tasic_metadata_vignette)
}
\description{
A DataFrame with metadata derived from Tasic case study (explain how).
}
\keyword{datasets}
