% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test.to.ladder.idx}
\alias{test.to.ladder.idx}
\title{The index of the region in the ladder regions, given its index in the test regions}
\usage{
test.to.ladder.idx(idx)
}
\description{
The index of the region in the ladder regions, given its index in the test regions
}

