% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTable.R
\name{topTable}
\alias{topTable}
\title{Test function to obtain a top list of transcripts that are differentially used
in the contrast of interest}
\usage{
topTable(models, contrast, plot = FALSE)
}
\arguments{
\item{models}{A list of objects of the `StatModel` class as obtained through the
`fitQB` function of the `qbDTU` package.}

\item{contrast}{A `vector` or `matrix` that specifies the contrast of interest.}

\item{plot}{`boolean(1)` Logical, defaults to FALSE. If set to TRUE,
a plot of the histogram of the empirical z-scores and the standard normal
distribution.}
}
\value{
A `Dataframe` displaying the significance of DTU for each transcript.
     A `List` of `Dataframes` will be returned if multiple contrasts of interest
     were provided in `contrast`.
}
\description{
Test function to assesss differentially transcript usage (DTU)
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
