% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{train.factor.Rle.to.test}
\alias{train.factor.Rle.to.test}
\title{Convert a factor-Rle that is indexed by the training ranges to a factor-Rle
that is indexed by the test ranges, with NA filling in the missing values}
\usage{
train.factor.Rle.to.test(x)
}
\description{
Convert a factor-Rle that is indexed by the training ranges to a factor-Rle
that is indexed by the test ranges, with NA filling in the missing values
}

