% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{train.to.test.idx}
\alias{train.to.test.idx}
\title{The index of the region in the test regions, given its index in the training regions}
\usage{
train.to.test.idx(idx)
}
\description{
The index of the region in the test regions, given its index in the training regions
}

