% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/fitDTU.R
\name{fitDTU}
\alias{fitDTU}
\alias{fitDTU,SummarizedExperiment-method}
\title{fitDTU}
\usage{
fitDTU(object, ...)

\S4method{fitDTU}{SummarizedExperiment}(
  object,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam(),
  verbose = TRUE
)
}
\arguments{
\item{object}{A `SummarizedExperiment` instance}

\item{...}{parameters including:}

\item{parallel}{Logical, defaults to FALSE. Set to TRUE if you want to
parallellize the fitting.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{bpparam} in \code{BiocParallel} package for details.}

\item{verbose}{Logical, should progress be printed?}
}
\value{
An updated `SummarizedExperiment` instance. The instance now includes
a new list of models ("fitDTUModels") in its rowData slot, which can be accessed
by rowData(object)[["fitDTUModels"]].
}
\description{
Parameter estimation of quasi-binomial models.
}
\examples{
data(sumExp_vignette, package = "satuRn")
data(Tasic_metadata_vignette, package = "satuRn")
Tasic_metadata_vignette$group <- paste(Tasic_metadata_vignette$brain_region, Tasic_metadata_vignette$cluster, sep = ".")
sumExp <- fitDTU(
   object = sumExp_vignette,
   parallel = TRUE,
   BPPARAM = BiocParallel::bpparam(),
   verbose = TRUE)

}
\author{
Jeroen Gilis
}
