% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQB.R
\name{fitQB,SummarizedExperiment-method}
\alias{fitQB,SummarizedExperiment-method}
\title{Function to fit quasi-binomial models}
\usage{
\S4method{fitQB}{SummarizedExperiment}(
  object,
  parallel = FALSE,
  BPPARAM = BiocParallel::bpparam(),
  verbose = TRUE
)
}
\arguments{
\item{object}{`SummarizedExperiment` instance}

\item{modelColumnName}{`character` to indicate the variable name that is used
to store the fitQB models in the rowData of the SummarizedExperiment
instance or of the assay of the Features instance. Default is "fitQBModels".}
}
\value{
A list of objects of the `StatModel` class.
}
\description{
Parameter estimation of quasi-binomial models.
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
