% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDTU.R
\name{plotDTU}
\alias{plotDTU}
\title{Plot function to visualize differential transcript usage (DTU)}
\usage{
plotDTU(
  object,
  contrast,
  groups,
  coefficients,
  summaryStat = "model",
  transcripts = NULL,
  genes = NULL,
  top.n = 6
)
}
\arguments{
\item{object}{A `SummarizedExperiment` containing the models and results
of the DTU analysis as obtained by the `fitDTU` and `testDTU` function from
this `satuRn` package, respectively.}

\item{contrast}{Specifies the specific contrast for which the visualization should be
returned.}

\item{summaryStat}{Which summaryStatistic for the relative usage of the transcript
should be displayed. `Character` or `character vector`, must be any of following
summary statistics; model (default), mean or median.}

\item{transcripts}{A `character` or `character vector` of transcript IDs,
to specify which transcripts should be visualized. Can be used together with
`genes`. If not specified, `plotDTU` will check if the `genes` slot is
specified.}

\item{genes}{A single `character` or `character vector` of gene IDs,
to specify the genes for which the individual transcripts should be visualized.
Can be used together with `transcripts`. If not specified, `plotDTU` will check
if the `transcripts` slot is specified.}

\item{top.n}{A `numeric` value. If neither `transcripts` nor `genes` was
was specified, this argument leads to the visualization of the `n` most
significant DTU transcripts in the contrast. Defaults to 6 transcripts}
}
\value{
A ggplot object that can be directly displayed in the current R session or stored in a list.
}
\description{
Plot function to visualize differential transcript usage (DTU)
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
