% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tasic_metadata_vignette}
\alias{Tasic_metadata_vignette}
\title{Metadata associated with the expression matrix `Tasic_counts_vignette`. 
See `?Tasic_counts_vignette` for more information on the dataset.}
\format{
An object of class \code{data.frame} with 60 rows and 3 columns.
}
\usage{
data(Tasic_metadata_vignette)
}
\description{
Metadata associated with the expression matrix `Tasic_counts_vignette`. 
See `?Tasic_counts_vignette` for more information on the dataset.
}
\keyword{datasets}
