% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDTU.R
\name{testDTU}
\alias{testDTU}
\title{Test function to obtain a top list of transcripts that are differentially used
in the contrast of interest}
\usage{
testDTU(object, contrasts, plot = FALSE, sort = FALSE)
}
\arguments{
\item{plot}{`boolean(1)` Logical, defaults to FALSE. If set to TRUE,
a plot of the histogram of the empirical z-scores and the standard normal
distribution.}

\item{sort}{`boolean(1)` Logical, defaults to FALSE. If set to TRUE, the output}

\item{models}{A `SummarizedExperiment` containing a list of objects of the `StatModel`
class as obtained by the `fitDTU` function of the `satuRn` package.}

\item{contrast}{A `vector` or `matrix` that specifies the contrast of interest.}
}
\value{
An updated `SummarizedExperiment` that contains the `Dataframes` displaying 
     the significance of DTU for each transcript in each contrast of interest.
}
\description{
Test function to assesss differentially transcript usage (DTU)
}
\examples{
#TODO

}
\author{
Jeroen Gilis
}
