% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDTU.R
\name{testDTU}
\alias{testDTU}
\title{Test function to obtain a top list of transcripts that are differentially used
in the contrast of interest}
\usage{
testDTU(object, contrasts, plot = FALSE, sort = FALSE)
}
\arguments{
\item{object}{A `SummarizedExperiment` instance containing a list of objects of
the `StatModel` class as obtained by the `fitDTU` function of the `satuRn` package.}

\item{contrasts}{`numeric` matrix specifying one or more contrasts of
the linear model coefficients to be tested.
The rownames of the matrix should be equal to the names
of parameters of the model that are involved in the contrast.
The column names of the matrix will be used to construct names to store
the results in the rowData of the SummarizedExperiment.}

\item{plot}{`boolean(1)` Logical, defaults to FALSE. If set to TRUE,
a plot of the histogram of the empirical z-scores and the standard normal
distribution will bee displayed.}

\item{sort}{`boolean(1)` Logical, defaults to FALSE. If set to TRUE, the output}
}
\value{
An updated `SummarizedExperiment` that contains the `Dataframes` displaying
     the significance of DTU for each transcript in each contrast of interest.
}
\description{
Function to test for differential transcript usage (DTU)
}
\examples{
data(sumExp_vignette, package = "satuRn")
data(Tasic_metadata_vignette, package = "satuRn")
Tasic_metadata_vignette$group <- paste(Tasic_metadata_vignette$brain_region, Tasic_metadata_vignette$cluster, sep = ".")
sumExp <- fitDTU(
   object = sumExp_vignette,
   parallel = TRUE,
   BPPARAM = BiocParallel::bpparam(),
   verbose = TRUE)
group <- as.factor(Tasic_metadata_vignette$group)
design <- model.matrix(~ 0 + group)
colnames(design) <- levels(group)
L <- matrix(0, ncol = 2, nrow = ncol(design))
rownames(L) <- colnames(design)
colnames(L) <- c("Contrast1", "Contrast2")
L[c("VISp.L5_IT_VISp_Hsd11b1_Endou", "ALM.L5_IT_ALM_Tnc"), 1] <- c(1, -1)
L[c("VISp.L5_IT_VISp_Hsd11b1_Endou", "ALM.L5_IT_ALM_Tmem163_Dmrtb1"), 2] <- c(1, -1)

sumExp <- testDTU(object = sumExp, contrasts = L, plot = TRUE, sort = TRUE)

}
\author{
Jeroen Gilis
}
